Rename ckb_chain::consume_unverified to ckb_chain::verify
