+++ chain/src/tests/find_fork.rs
@@ -1,5 +1,5 @@
-use crate::consume_unverified::ConsumeUnverifiedBlockProcessor;
 use crate::utils::forkchanges::ForkChanges;
+use crate::verify::ConsumeUnverifiedBlockProcessor;
 use crate::{start_chain_services, UnverifiedBlock};
 use ckb_chain_spec::consensus::{Consensus, ProposalWindow};
 use ckb_proposal_table::ProposalTable;
